requirements.txt
flask
requests
flask-RESTful
boto3
flask_sqlalchemy
Pillow
cmake
dlib
ipython
matplotlib
numpy
opencv-python
scipy
tensorboard
torch
torchvision
tqdm
flask_cors
ninja
mediapipe